Previous ucore's broken... Now replaced and updated...

+++ src/ucore/kern/driver/clock.h
@@ -1,7 +1,9 @@
 #ifndef __KERN_DRIVER_CLOCK_H__
 #define __KERN_DRIVER_CLOCK_H__
 
-extern volatile unsigned long ticks;
+#include <types.h>
+
+extern volatile size_t ticks;
 
 void clock_init(void);
