+++ src/ucore/kern/mm/swap.h
@@ -12,21 +12,21 @@
  *           24 bits            7 bits    1 bit
  * */
 
-#define MAX_SWAP_OFFSET_LIMIT					(1 << 24)
+#define MAX_SWAP_OFFSET_LIMIT                   (1 << 24)
 
-extern unsigned int max_swap_offset;
+extern size_t max_swap_offset;
 
 /* *
  * swap_offset - takes a swap_entry (saved in pte), and returns
  * the corresponding offset in swap mem_map.
  * */
-#define swap_offset(entry) ({								\
-	unsigned int __offset = (entry >> 8);					\
-	if (!(__offset > 0 && __offset < max_swap_offset)) {	\
-		panic("invalid swap_entry_t = %08x.\n", entry);		\
-	}														\
-	__offset;												\
-})
+#define swap_offset(entry) ({                                       \
+            size_t __offset = (entry >> 8);                         \
+            if (!(__offset > 0 && __offset < max_swap_offset)) {    \
+                panic("invalid swap_entry_t = %08x.\n", entry);     \
+            }                                                       \
+            __offset;                                               \
+        })
 
 void swap_init(void);
 bool try_free_pages(size_t n);
